requirements.txt
mysql-connector-python
configparser
datetime